refactor: 컴포넌트별 하위 폴더 구조로 재편 및 import 경로 수정
