+++ src/feature/CardRegister/components/fields/CvcField/CvcField.tsx
@@ -1,20 +1,20 @@
 import Input from '@/common/components/Input/Input';
 import styled from 'styled-components';
 import type {CardRegisterInputProps} from '@/feature/CardRegister/types/cardRegisterInputProps';
-import InputContainer from './InputContainer';
-import {FieldErrorMessage, FieldLayout} from '../styles/inputFieldStyles';
+import InputContainer from '../InputContainer/InputContainer';
+import {FieldErrorMessage, FieldLayout} from '../../styles/inputFieldStyles';
 
-type PasswordFieldProps = {
+type CvcFieldProps = {
   inputProps: CardRegisterInputProps;
   errorMessage: string;
 };
 
-const PasswordField = ({inputProps, errorMessage}: PasswordFieldProps) => {
+const CvcField = ({inputProps, errorMessage}: CvcFieldProps) => {
   return (
-    <InputContainer title='비밀번호를 입력해 주세요' description='앞의 2자리를 입력해주세요' label='비밀번호 앞 2자리'>
+    <InputContainer title='CVC 번호를 입력해 주세요' label='CVC'>
       <FieldLayout>
         <InputWrapper>
-          <PasswordInput {...inputProps} strokeMode={errorMessage ? 'error' : 'default'} />
+          <CvcInput {...inputProps} strokeMode={errorMessage ? 'error' : 'default'} />
         </InputWrapper>
         <FieldErrorMessage>{errorMessage}</FieldErrorMessage>
       </FieldLayout>
@@ -25,12 +25,13 @@ const PasswordField = ({inputProps, errorMessage}: PasswordFieldProps) => {
 const InputWrapper = styled.div`
   display: grid;
   grid-template-columns: minmax(0, 1fr);
+  gap: 14px;
 `;
 
-const PasswordInput = styled(Input)`
+const CvcInput = styled(Input)`
   box-sizing: border-box;
   width: 100%;
   height: 32px;
 `;
 
-export default PasswordField;
+export default CvcField;
