+++ src/feature/CardRegister/components/fields/NumberField/NumberField.tsx
@@ -3,8 +3,8 @@ import styled from 'styled-components';
 import Input from '@/common/components/Input/Input';
 import type {CardRegisterInputProps} from '@/feature/CardRegister/types/cardRegisterInputProps';
 import {useInputFocusMove} from '@/feature/CardRegister/hooks/useInputFocusMove';
-import InputContainer from './InputContainer';
-import {FieldErrorMessage, FieldLayout} from '../styles/inputFieldStyles';
+import InputContainer from '../InputContainer/InputContainer';
+import {FieldErrorMessage, FieldLayout} from '../../styles/inputFieldStyles';
 
 type NumberFieldProps = {
   inputProps: CardRegisterInputProps[];
