+++ src/feature/CardRegister/components/preview/CardExpiryDateDisplay/CardExpiryDateDisplay.stories.tsx
