+++ src/feature/CardRegister/components/preview/CardExpiryDateDisplay/CardExpiryDateDisplay.tsx
