+++ src/feature/CardRegister/components/preview/CardNumberDisplay/CardNumberDisplay.tsx
