+++ src/feature/CardRegister/components/sections/CardPreviewSection/CardPreviewSection.tsx
