+++ src/feature/CardRegister/components/sections/InfoInputSection/InfoInputSection.tsx
